Revert unnecessary changes
