+++ web/app/templates/authenticated/new/doc.hbs
@@ -1,3 +1,3 @@
-{{page-title (concat "Create Your " @model.name)}}
+{{page-title (concat "Create Your " @model)}}
 
-<New::DocForm @docType={{@model.name}} />
+<New::DocForm @docType={{@model}} />
